+++ crates/miden-testing/src/kernel_tests/tx/test_tx.rs
@@ -42,7 +42,7 @@ use miden_objects::{
 };
 use miden_tx::{
     ExecutionOptions, ScriptMastForestStore, TransactionExecutor, TransactionExecutorError,
-    TransactionExecutorHost, TransactionMastStore,
+    TransactionExecutorHost, TransactionMastStore, auth::UnreachableAuth,
 };
 use vm_processor::{AdviceInputs, Process, crypto::RpoRandomCoin};
 
@@ -124,7 +124,7 @@ fn consuming_note_created_in_future_block_fails() -> anyhow::Result<()> {
     let tx_context = mock_chain.build_tx_context(account.id(), &[], &[])?.build()?;
     let source_manager = tx_context.source_manager();
 
-    let tx_executor = TransactionExecutor::new(&tx_context, None);
+    let tx_executor = TransactionExecutor::<'_, '_, _, UnreachableAuth>::new(&tx_context, None);
     // Try to execute with block_ref==1
     let error = tx_executor.execute_transaction(
         account.id(),
@@ -920,7 +920,7 @@ fn advice_inputs_from_transaction_witness_are_sufficient_to_reexecute_transactio
     let mast_store = Arc::new(TransactionMastStore::new());
     mast_store.load_account_code(tx_inputs.account().code());
 
-    let mut host = TransactionExecutorHost::new(
+    let mut host = TransactionExecutorHost::<'_, '_, _, UnreachableAuth>::new(
         &tx_inputs.account().into(),
         &mut advice_inputs,
         mast_store.as_ref(),
@@ -1319,7 +1319,7 @@ fn execute_tx_view_script() -> anyhow::Result<()> {
     let block_ref = tx_context.tx_inputs().block_header().block_num();
     let advice_inputs = tx_context.tx_args().advice_inputs().clone();
 
-    let executor = TransactionExecutor::new(&tx_context, None);
+    let executor = TransactionExecutor::<'_, '_, _, UnreachableAuth>::new(&tx_context, None);
 
     let stack_outputs = executor.execute_tx_view_script(
         account_id,