+++ crates/miden-tx/src/auth/mod.rs
@@ -1,4 +1,4 @@
 mod tx_authenticator;
-pub use tx_authenticator::{BasicAuthenticator, TransactionAuthenticator};
+pub use tx_authenticator::{BasicAuthenticator, TransactionAuthenticator, UnreachableAuth};
 
 pub mod signatures;